Moved non-table code out of tables/table_common.js (by fluke likely, it was only two specific panels using this code still)
